Refactor ActionButton test from enzyme to react Remove snapshot
